move testutil and maputil into util
