+++ pkg/app/load_opts.go
@@ -3,8 +3,8 @@ package app
 import (
 	"gopkg.in/yaml.v3"
 
-	"github.com/helmfile/helmfile/pkg/maputil"
 	"github.com/helmfile/helmfile/pkg/state"
+	"github.com/helmfile/helmfile/pkg/util"
 )
 
 type LoadOpts struct {
@@ -22,7 +22,7 @@ type LoadOpts struct {
 }
 
 func (o LoadOpts) DeepCopy() LoadOpts {
-	bytes, err := maputil.YamlMarshal(o)
+	bytes, err := util.YamlMarshal(o)
 	if err != nil {
 		panic(err)
 	}
