+++ pkg/config/config.go
@@ -3,7 +3,7 @@ package config
 import (
 	"strings"
 
-	"github.com/helmfile/helmfile/pkg/maputil"
+	"github.com/helmfile/helmfile/pkg/util"
 )
 
 func NewCLIConfigImpl(g *GlobalImpl) error {
@@ -14,10 +14,10 @@ func NewCLIConfigImpl(g *GlobalImpl) error {
 			ops := strings.Split(optsSet[i], ",")
 			for j := range ops {
 				op := strings.SplitN(ops[j], "=", 2)
-				k := maputil.ParseKey(op[0])
+				k := util.ParseKey(op[0])
 				v := op[1]
 
-				maputil.Set(set, k, v)
+				util.Set(set, k, v)
 			}
 		}
 		g.SetSet(set)
