+++ pkg/state/chart_dependency.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/helmfile/helmfile/pkg/app/version"
 	"github.com/helmfile/helmfile/pkg/helmexec"
-	"github.com/helmfile/helmfile/pkg/maputil"
+	"github.com/helmfile/helmfile/pkg/util"
 )
 
 type ChartMeta struct {
@@ -300,7 +300,7 @@ func (m *chartDependencyManager) updateHelm3(shell helmexec.DependencyUpdater, w
 	chartMetaContent := fmt.Sprintf("name: %s\nversion: 1.0.0\napiVersion: v2\n", m.Name)
 
 	// Generate `requirements.yaml` of the temporary local chart from the helmfile state
-	reqsContent, err := maputil.YamlMarshal(unresolved.ToChartRequirements())
+	reqsContent, err := util.YamlMarshal(unresolved.ToChartRequirements())
 	if err != nil {
 		return nil, err
 	}
@@ -318,7 +318,7 @@ func (m *chartDependencyManager) updateHelm2(shell helmexec.DependencyUpdater, w
 	}
 
 	// Generate `requirements.yaml` of the temporary local chart from the helmfile state
-	reqsContent, err := maputil.YamlMarshal(unresolved.ToChartRequirements())
+	reqsContent, err := util.YamlMarshal(unresolved.ToChartRequirements())
 	if err != nil {
 		return nil, err
 	}
@@ -384,7 +384,7 @@ func (m *chartDependencyManager) doUpdate(chartLockFile string, unresolved *Unre
 
 	lockedReqs.Version = version.Version
 
-	updatedLockFileContent, err = maputil.YamlMarshal(lockedReqs)
+	updatedLockFileContent, err = util.YamlMarshal(lockedReqs)
 
 	if err != nil {
 		return nil, err
