+++ pkg/state/create.go
@@ -14,8 +14,8 @@ import (
 	"github.com/helmfile/helmfile/pkg/environment"
 	"github.com/helmfile/helmfile/pkg/filesystem"
 	"github.com/helmfile/helmfile/pkg/helmexec"
-	"github.com/helmfile/helmfile/pkg/maputil"
 	"github.com/helmfile/helmfile/pkg/remote"
+	"github.com/helmfile/helmfile/pkg/util"
 )
 
 const (
@@ -313,7 +313,7 @@ func (c *StateCreator) scatterGatherEnvSecretFiles(st *HelmState, envSecretFiles
 				// All the nested map key should be string. Otherwise we get strange errors due to that
 				// mergo or reflect is unable to merge map[interface{}]interface{} with map[string]interface{} or vice versa.
 				// See https://github.com/roboll/helmfile/issues/677
-				vals, err := maputil.CastKeysToStrings(m)
+				vals, err := util.CastKeysToStrings(m)
 				if err != nil {
 					results <- secretResult{secret.id, nil, fmt.Errorf("failed to load environment secrets file \"%s\": %v", secret.path, err), secret.path}
 					continue
