+++ pkg/state/release.go
@@ -5,8 +5,8 @@ import (
 
 	"gopkg.in/yaml.v3"
 
-	"github.com/helmfile/helmfile/pkg/maputil"
 	"github.com/helmfile/helmfile/pkg/tmpl"
+	"github.com/helmfile/helmfile/pkg/util"
 )
 
 func (r ReleaseSpec) ExecuteTemplateExpressions(renderer *tmpl.FileRenderer) (*ReleaseSpec, error) {
@@ -99,7 +99,7 @@ func (r ReleaseSpec) ExecuteTemplateExpressions(renderer *tmpl.FileRenderer) (*R
 		for i, t := range result.ValuesTemplate {
 			switch ts := t.(type) {
 			case map[interface{}]interface{}, map[string]interface{}:
-				serialized, err := maputil.YamlMarshal(ts)
+				serialized, err := util.YamlMarshal(ts)
 				if err != nil {
 					return nil, fmt.Errorf("failed executing template expressions in release \"%s\".values[%d] = \"%v\": %v", r.Name, i, ts, err)
 				}
@@ -202,7 +202,7 @@ func (r ReleaseSpec) ExecuteTemplateExpressions(renderer *tmpl.FileRenderer) (*R
 }
 
 func (r ReleaseSpec) Clone() (*ReleaseSpec, error) {
-	serialized, err := maputil.YamlMarshal(r)
+	serialized, err := util.YamlMarshal(r)
 	if err != nil {
 		return nil, fmt.Errorf("failed cloning release \"%s\": %v", r.Name, err)
 	}
