+++ pkg/util/test_test.go
@@ -1,4 +1,4 @@
-package testutil
+package util
 
 import (
 	"fmt"
