+++ pkg/util/yaml_test.go
@@ -1,4 +1,4 @@
-package maputil
+package util
 
 import (
 	"testing"
